LayoutTests/http/tests/xmlhttprequest/navigation-should-abort-expected.txt

CONSOLE MESSAGE: line 14: PASS: Expected 'abort', got 'abort'.
This page should load, and the XMLHttpRequest from the previous page should have aborted and fired a 'PASS' message to the console.
